moved SearchScholarIterator to separate file to improve readability
